refactor: rename to avoid confusion with actual wallet
