+++ src/context/walletContext.tsx
@@ -2,16 +2,16 @@ import React, { PropsWithChildren, useState } from 'react';
 import { Core } from '@walletconnect/core';
 import { ICore } from '@walletconnect/types';
 import { IWeb3Wallet } from '@walletconnect/web3wallet';
-// @ts-expect-error - env is a virtualised module via Babel config.
+// @ts-expect-error - env is a virtualized module via Babel config.
 import { ENV_PROJECT_ID, ENV_RELAY_URL } from '@env';
 
 import { createCtx } from '.';
-import { createWeb3Wallet } from '../utils/wallet';
+import { createWeb3Wallet } from '../utils/web3';
 
 // State variables only
 interface WalletContextState {
-  core: ICore;
-  wallet?: IWeb3Wallet;
+  web3Core: ICore;
+  web3Wallet?: IWeb3Wallet;
 }
 
 // This interface differentiates from State
@@ -22,7 +22,7 @@ interface WalletContext extends WalletContextState {
 }
 
 const INITIAL_STATE: WalletContextState = {
-  core: new Core({
+  web3Core: new Core({
     // @notice: If you want the debugger / logs
     // logger: 'debug',
     projectId: ENV_PROJECT_ID,
@@ -37,8 +37,8 @@ export const WalletProvider = ({ children }: PropsWithChildren) => {
   const [state, setState] = useState<WalletContextState>(INITIAL_STATE);
 
   const initWallet = async () =>
-    await createWeb3Wallet(state.core).then(wallet =>
-      setState(prevState => ({ ...prevState, wallet })),
+    await createWeb3Wallet(state.web3Core).then(web3Wallet =>
+      setState(prevState => ({ ...prevState, web3Wallet })),
     );
 
   return (
